DF-7369 MMAnotherService now can receive snapshots
Former-commit-id: b58dea51d51c90e8b937a020bffad51c45dc7a7f

+++ Tools/ResourceEditor/Classes/Qt/DeviceInfo/DeviceInfo/MemProfWidget.cpp
@@ -121,7 +121,7 @@ void MemProfWidget::StatArrived(size_t /*itemCount*/)
     UpdatePlot(stat);
 }
 
-void MemProfWidget::SnapshotArrived(size_t sizeTotal, size_t sizeRecv)
+void MemProfWidget::SnapshotArrived(unsigned int sizeTotal, unsigned int sizeRecv)
 {
     if (sizeTotal > 0 && sizeRecv > 0)
     {
